requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
Django==3.2.9
docker==4.2.0
Flask==1.1.1
Flask-Migrate==4.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==1.0.1
httplib2==0.20.4
Jinja2==3.0.3
lxml==4.8.0
PyMySQL==0.9.3
python-nmap==0.7.1
python-utils==3.3.3
python3-nmap==1.4.1
redis==3.4.1
requests==2.23.0
SQLAlchemy==1.3.15
urllib3==1.25.8
wappalyzer==1.0
Werkzeug==1.0.0
WTForms==3.0.1
wxPython==4.2.0
xlwt==1.3.0
